Fix: consertando rotas autenticadas
