+++ src/main.jsx
@@ -4,7 +4,7 @@ import "./index.css";
 
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import { CatalogoClinicas } from "./pages/CatalogoClinicas/CatalogoClinicas.jsx";
-import InitialPage from "./pages/InitialPage/InitialPage.jsx";
+import { InitialPage } from "./pages/InitialPage/InitialPage.jsx";
 import { CatalogoEventos } from "./pages/CatalogoEventos/CatalogoEventos.jsx";
 import { Home } from "./pages/Home/Home.jsx";
 import { SobreNos } from "./pages/SobreNos/SobreNos.jsx";
@@ -15,9 +15,10 @@ import { SelectBall } from "./pages/SelectBall/SelectBall.jsx";
 import { Config } from "./pages/Config/Config.jsx";
 import { ListagemSquad } from "./pages/ListagemSquad/ListagemSquad.jsx";
 import { Chat } from "./pages/Chat/Chat.jsx";
-import MainLayout from "./MainLayout.jsx";
+import { MainLayout } from "./MainLayout.jsx";
 import { Login } from "./pages/Login/Login.jsx";
 import { App } from "./App.jsx";
+import { AuthLayer } from "./AuthLayer.jsx";
 
 const router = createBrowserRouter([
   {
@@ -33,50 +34,56 @@ const router = createBrowserRouter([
     element: <Login />,
   },
   {
-    path: "/ball",
-    element: <SelectBall />,
-  },
-  {
-    path: "/trilhas",
-    element: <Trilhas />,
-  },
-  {
     path: "/",
-    element: <MainLayout />,
+    element: <AuthLayer />,
     children: [
       {
-        path: "catalogo-clinicas",
-        element: <CatalogoClinicas />,
+        path: "/ball",
+        element: <SelectBall />,
       },
       {
-        path: "home",
-        element: <Home />,
+        path: "/trilhas",
+        element: <Trilhas />,
       },
       {
-        path: "sobre-nos",
-        element: <SobreNos />,
+        path: "/",
+        element: <MainLayout />,
+        children: [
+          {
+            path: "catalogo-clinicas",
+            element: <CatalogoClinicas />,
+          },
+          {
+            path: "home",
+            element: <Home />,
+          },
+          {
+            path: "sobre-nos",
+            element: <SobreNos />,
+          },
+          {
+            path: "squads",
+            element: <Squads />,
+          },
+          {
+            path: "config",
+            element: <Config />,
+          },
+          {
+            path: "listagem-squad",
+            element: <ListagemSquad />,
+          },
+          {
+            path: "chat",
+            element: <Chat />,
+          },
+          {
+            path: "/catalogo-eventos",
+            element: <CatalogoEventos />,
+          },
+        ],
       },
-      {
-        path: "squads",
-        element: <Squads />,
-      },
-      {
-        path: "config",
-        element: <Config />,
-      },
-      {
-        path: "listagem-squad",
-        element: <ListagemSquad />,
-      },
-      {
-        path: "chat",
-        element: <Chat />,
-      },
-      {
-        path: "/catalogo-eventos",
-        element: <CatalogoEventos />,
-      },
-    ],
+    ]
   },
 ]);
