+++ src/pages/Login/Login.jsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import img from "../../../public/icons/login.png";
 import logo from "../../../public/icons/logo.png";
 import styles from "./Login.module.css";
@@ -13,6 +13,7 @@ import { httpClient } from "../../services/httpClient";
 import { useNavigate } from "react-router-dom";
 
 export const Login = () => {
+  const navigate = useNavigate();
   const [isLoading, setIsLoading] = useState(false)
   const [isDisabled, setIsDisabled] = useState(true);
   const [erros, setErros] = useState([]);
@@ -20,7 +21,6 @@ export const Login = () => {
     email: "",
     senha: "",
   });
-  const navigate = useNavigate();
 
   const handleInputChange = (e) => {
     const { name, value } = e.target;
@@ -51,6 +51,16 @@ export const Login = () => {
     }
     setIsLoading(false);
   }
+
+  const handleIsAutenticado = () => {
+    const token = localStorage.getItem("jwt_token");
+
+    if (token) navigate("/home");
+  }
+
+  useEffect(() => {
+    handleIsAutenticado();
+  }, []);
 
   return (
     <Spin spinning={isLoading}>
